chore(cleanup): ruff auto-fix de 44 errores pre-existentes en main
Main venía con 44 ruff errors heredados (bloqueaban lint en cualquier
PR que tocara apps/api/*). Todos cosméticos y fixeables con
`ruff check . --fix --unsafe-fixes`:

- **RUF059 × 41**: tuples unpacked con variables no usadas en tests.
  Rename: `t, _u, auth = await _admin(...)` → `_t, _u, auth = ...`
  (y similares en 13 archivos de tests). Cero cambio de lógica —
  solo el prefijo `_` que le dice al linter "intencional".
- **RUF046 × 2**: `int(round(x))` en `operational_reports.py:86` y
  `xlsx_task_parser.py:133`. En Python 3, `round(float)` sin ndigits
  ya devuelve `int` — el `int()` wrapper es redundante.
- **UP047 × 1**: `workers/db.py` — `def run_async(coro: ...) -> T`
  con TypeVar module-level pasa a `def run_async[T](coro: ...)`
  usando PEP 695 (sintaxis nueva de Python 3.12). Se elimina el
  `T = TypeVar("T")` ahora huérfano y el `from typing import TypeVar`.

Run final: `ruff check .` → "All checks passed!".

Este commit se hace en la branch de US-069 porque la lint-gate de CI
bloqueaba el merge del PR #143 con estos errores heredados —
técnicamente no son bugs de la US pero sin limpiarlos el PR no pasa.

+++ apps/api/app/services/xlsx_task_parser.py
@@ -130,7 +130,7 @@ def _coerce_progress(v: object) -> int:
         return 0
     if 0 <= n <= 1.0 and "." in s:
         n = n * 100
-    return max(0, min(100, int(round(n))))
+    return max(0, min(100, round(n)))
 
 
 def _coerce_bool(v: object) -> bool:
